requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database - PostgreSQL
asyncpg==0.29.0

# Redis
redis[hiredis]==5.2.0

# Security
pwdlib[argon2]==0.2.1
python-jose[cryptography]==3.3.0
PyJWT==2.9.0

# HTTP Client
httpx==0.27.2

# Rate Limiting
slowapi==0.1.9

# Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Environment
python-dotenv==1.0.1
